Move DefaultPrompt and PromptMethod enums to constants.ts
