+++ src/wagtail_ai/static_src/constants.ts
@@ -1,0 +1,11 @@
+export enum DefaultPrompt {
+  CORRECTION = 1,
+  COMPLETION = 2,
+  DESCRIPTION = 3,
+  TITLE = 4,
+}
+
+export enum PromptMethod {
+  APPEND = 'append',
+  REPLACE = 'replace',
+}
